Atualização da arquitetura para a rota de edição de user
