+++ Lyrics_Lab/Controllers/UserController.cs
@@ -1,0 +1,26 @@
+﻿using Lyrics_Lab.DTOs;
+using Lyrics_Lab.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Lyrics_Lab.Controllers
+{
+    [Route("api/[controller]")]
+    [ApiController]
+    [Authorize]
+    public class UserController : ControllerBase
+    {
+        private readonly IUserService _userService;
+
+        public UserController(IUserService userService)
+        {
+            _userService = userService;
+        }
+
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UpdateUser(int id, [FromBody] UpdateUserDto updateUserDto)
+        {
+            return await _userService.UpdateUser(id, updateUserDto, User);
+        }
+    }
+}
