+++ Lyrics_Lab/Services/Interfaces/IUserService.cs
@@ -1,6 +1,7 @@
 ﻿using Lyrics_Lab.DTOs;
 using Lyrics_Lab.Models;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace Lyrics_Lab.Services.Interfaces
 {
@@ -10,5 +11,6 @@ namespace Lyrics_Lab.Services.Interfaces
         IActionResult Login(LoginDto dto);
         IActionResult GetUser(string jwt);
         IActionResult Logout(HttpResponse response);
+        Task<IActionResult> UpdateUser(int id, UpdateUserDto updateUserDto, ClaimsPrincipal userClaims);
     }
 }
